+++ CHANGELOG.md
@@ -2,11 +2,17 @@
 
 ## Current Version
 
+v1.1.0
+
+- Retarget to include .NET 6.0, .NET 7.0
+- Async APIs
+- Minor breaking changes (including ```GoogleMapsAddress```, ```GoogleMapsCoordinates```, ```GoogleMapsTimestamp``` return object)
+
+## Previous Versions
+
 v1.0.1
 
 - LocalTimestamp API
-
-## Previous Versions
 
 v1.0.0